requirements.txt
python-telegram-bot
aiofiles
shapely
pydantic